miniFAS/requirements.txt

psutil

easydict==1.13

Flask==3.0.0

Flask-Cors==4.0.1

keras==3.3.3

matplotlib==3.8.2

mtcnn==0.1.1

numpy==1.26.2

onnx==1.16.0

onnx-tf==1.10.0

onnxruntime==1.16.0

onnxscript==0.1.0.dev20240329

opencv-python==4.8.1.78

pandas==2.1.4

Pillow==10.1.0

scikit-image==0.22.0

scikit-learn==1.3.2

scipy==1.11.4

tensorboard==2.16.2

tensorboard-data-server==0.7.2

tensorboard-plugin-wit==1.8.1

tensorboardX==2.6.2.2

tensorflow-probability==0.15.0

tensorflowjs==4.20.0

tf_keras==2.16.0

tf2onnx==1.16.1

torch==2.1.2

torchsummary==1.5.1

torchvision==0.16.2

tqdm==4.66.1

triton==2.1.0

